Move background unit tests to their own module
