+++ pyproject.toml
@@ -5,6 +5,7 @@ requires = ["setuptools>=30.3.0", "wheel", "setuptools_scm"]
 testpaths = [
     "tests/integration_test",
     "tests/server_unit_test",
+    "tests/background_unit_test",
 ]
 
 [tool.coverage.run]
